move workspace components to core
